+++ src/solution/data_structures/stacks/mod.rs
@@ -1,1 +1,2 @@
 mod min_stack;
+mod parentheses;
